+++ src/Core/Activator.php
@@ -32,8 +32,16 @@ final class Activator {
 		Database::install();
 		Settings::compile();
 
+		add_filter( 'cron_schedules', array( Plugin::class, 'cronSchedules' ) );
+
 		if ( ! wp_next_scheduled( 'gt_performance_run_queue' ) ) {
 			wp_schedule_event( time() + MINUTE_IN_SECONDS, 'gtp_every_minute', 'gt_performance_run_queue' );
 		}
+
+		if ( ! wp_next_scheduled( 'gt_performance_verify_license' ) ) {
+			wp_schedule_event( time() + DAY_IN_SECONDS, 'gtp_weekly', 'gt_performance_verify_license' );
+		}
+
+		remove_filter( 'cron_schedules', array( Plugin::class, 'cronSchedules' ) );
 	}
 }